third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-lifetime.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the RTCPeerConnection lifetime.");

var dc = null;

function dataChannelMessage(m)
{
    testPassed("dataChannelMessage");
    finishJSTest();
}

function dataChannelOpen()
{
    testPassed("dataChannelOpen");
    dc.onmessage = dataChannelMessage;
    shouldNotThrow("dc.send('xyzzy');");
}

function createPeerConnectionAndDataChannel()
{
    // The Peer Connection object is leaked
    var pc = new webkitRTCPeerConnection({iceServers:[]}, null);
    dc = pc.createDataChannel("label");
    dc.onopen = dataChannelOpen;
}

var pcB = null;
var observationB = null;

// Test that the PeerConnection object is gc'd if close is called.
var pcA = new webkitRTCPeerConnection(null, null);
var observationA = internals.observeGC(pcA);
pcA.close();
pcA = null;
asyncGC(function() {
    shouldBeTrue('observationA.wasCollected');
    observationA = null;

    // Test that the PeerConnection object is not gc'd if close is not called.
    pcB = new webkitRTCPeerConnection(null, null);
    observationB = internals.observeGC(pcB);
    pcB = null;
    asyncGC(function() {
        shouldBeFalse('observationB.wasCollected');
        observationB = null;

        // This test times out if the Peer connection object is garbage collected.
        createPeerConnectionAndDataChannel();
        gc();
    });
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
